+++ examples/authorization/search/search_authorization_by_code.rb
@@ -1,6 +1,6 @@
-require_relative "../boot"
+require_relative "../../boot"
 
-# Authorization by code
+# Search: Authorization by code
 #
 #   You need to give:
 #     - authorization code
