+++ examples/checkout/search/search_transaction_by_reference.rb
@@ -1,4 +1,4 @@
-require_relative "../boot"
+require_relative "../../boot"
 
 # Transaction by reference
 #
