+++ examples/checkout/search/search_transactions_by_date.rb
@@ -1,4 +1,4 @@
-require_relative "../boot"
+require_relative "../../boot"
 require "active_support/all"
 
 # Transactions by date
