+++ examples/get_installments.rb
@@ -1,6 +1,6 @@
 require_relative "boot"
 
-# Installment
+# Get Installments
 #
 # Required AccountCredentials
 #
